+++ buildpacks/ruby/src/layers/shared.rs
@@ -78,34 +78,6 @@ mod tests {
     migrate_toml_chain! {TestMetadata}
 
     #[test]
-    fn test_restored_layer_action_returns_old_data() {
-        #[derive(Debug, Clone)]
-        struct AlwaysNoDiff {
-            value: String,
-        }
-        impl CacheDiff for AlwaysNoDiff {
-            fn diff(&self, _: &Self) -> Vec<String> {
-                vec![]
-            }
-        }
-
-        let old = AlwaysNoDiff {
-            value: "old".to_string(),
-        };
-        let now = AlwaysNoDiff {
-            value: "now".to_string(),
-        };
-
-        let result = restored_layer_action(&old, &now);
-        match result {
-            (RestoredLayerAction::KeepLayer, Meta::Data(data)) => {
-                assert_eq!(data.value, "old");
-            }
-            _ => panic!("Expected to keep layer"),
-        }
-    }
-
-    #[test]
     fn test_cached_layer_write_metadata_restored_layer_action() {
         let temp = tempfile::tempdir().unwrap();
         let context = temp_build_context::<RubyBuildpack>(temp.path());
